Move table builder to own file
